requirements.txt
Flask>=3.0,<3.1
Flask-SQLAlchemy>=3.1
SQLAlchemy>=2.0
Flask-Migrate>=4.0
psycopg[binary]>=3.1
python-dotenv>=1.0
alembic>=1.13
Flask-Login>=0.6
gunicorn>=21.2
